Move fetch and signed fetch into verify -> fetcher
